% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOptChr.R
\name{anyRcfileExists}
\alias{anyRcfileExists}
\title{anyRcfileExists}
\usage{
anyRcfileExists()
}
\value{
A logical.
}
\description{
Checks if any mecabrc file exists.
}
\details{
This is a helper function that checks if any mecabrc file exists
before initializing tagger.

'MeCab' expects a mecabrc file to be present;
if not, it will raise an error (without any message!).
}
