% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collScores.R
\name{collScores}
\alias{collScores}
\title{collScores}
\usage{
collScores(kekka, node, span)
}
\arguments{
\item{kekka}{Result of \code{\link[=collocate]{collocate()}}.}

\item{node}{Node word.}

\item{span}{Window span.}
}
\value{
A data frame.
}
\description{
Calculates T-score and MI-score according to the result of \code{\link[=collocate]{collocate()}}.
}
\examples{
if (anyRcfileExists()) {
  text_file <- system.file("samples/doc1.txt", package = "RMeCab")
  out <- collocate(text_file, "\u6570\u5b66")
  collScores(out, "\u6570\u5b66", 3)
}
}
