\name{coOccurrence}
\alias{coOccurrence}
\title{co-occurrence Matrix (Matrices) options}
\description{
  retrieve  co-occurrence term matrix }
\usage{
   coOccurrence(dtm, co = 1)
}
\arguments{
  \item{dtm}{a document-term matrix.}
  \item{co}{opitons 1 or 2: default being 1.}
}
\details{


}
\value{
    Returns the weighted textmatrix of the same size and format as the input matrix.
}
\references{
 \url{http://www.amazon.co.jp/dp/4627848412} 
}

\author{ Motohiro ISHIDA \email{ishida.motohiro@gmail.com} }
\examples{

}
\keyword{array}
