% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docDF.R
\name{docDF}
\alias{docDF}
\title{docDF}
\usage{
docDF(
  target,
  column = 0,
  type = 0,
  pos = NULL,
  minFreq = 1,
  N = 1,
  Genkei = 0,
  weight = "",
  nDF = 0,
  co = 0,
  dic = "",
  mecabrc = "",
  etc = ""
)
}
\arguments{
\item{target}{A file, directory, or a data.frame.}

\item{column}{Column number or name which include the text to analyze.}

\item{type}{Kind of tokens. \code{0} for character, \code{1} for term.
Defaults to \code{0}.}

\item{pos}{Parts of speech that should be extracted.
If \code{NULL}, all terms are extracted.}

\item{minFreq}{Minimum document frequency for filtering terms.
Terms that appear less than \code{minFreq} within a document are ignored.}

\item{N}{Unit of tokens. If \code{2}, counts bi-grams.}

\item{Genkei}{If \code{0}, counts genkei. Defaults to \code{0}.}

\item{weight}{Method to weight term frequencies.}

\item{nDF}{If \code{1}, N-grams are divided into columns.}

\item{co}{If \code{1}, returns co-occurrence matrix.}

\item{dic}{Path to a user dictionary file such as \code{ishida.dic}.}

\item{mecabrc}{Path to a mecab resource file.}

\item{etc}{Other options for 'MeCab' tagger.}
}
\value{
A data.frame is invisibly returned.
}
\description{
Counts tokens (characters, terms, or N-grams) within \code{target}.
\code{target} can be a file, directory, or a data.frame.
}
