% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docMatrix.R
\name{docMatrix}
\alias{docMatrix}
\title{docMatrix}
\usage{
docMatrix(
  mydir,
  pos = "Default",
  minFreq = 1,
  weight = "no",
  kigo = 0,
  co = 0,
  dic = "",
  mecabrc = "",
  etc = ""
)
}
\arguments{
\item{mydir}{A directory where text files are stored.}

\item{pos}{Parts of speech that should be extracted.
If \code{NULL}, all terms are extracted.}

\item{minFreq}{Minimum document frequency for filtering terms.
Terms that appear less than \code{minFreq} within a document are ignored.}

\item{weight}{Method to weight term frequencies.}

\item{kigo}{If \code{1}, \verb{[[TOTAL-TOKENS]]} includes number of symbols.
Defaults to \code{0} (does not count symbols).}

\item{co}{If \code{1}, returns co-occurrence matrix.}

\item{dic}{Path to a user dictionary file such as \code{ishida.dic}.}

\item{mecabrc}{Path to a mecab resource file.}

\item{etc}{Other options for 'MeCab' tagger.}
}
\value{
An integer matrix.
}
\description{
Creates a document-term matrix out of all files in a given directory.
Each cell of the matrix shows the actual frequency of each word.
}
