% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docMatrixDF.R
\name{docMatrixDF}
\alias{docMatrixDF}
\title{docMatrixDF}
\usage{
docMatrixDF(
  charVec = c("MeCab", "CaBoCha"),
  pos = "Default",
  minFreq = 1,
  weight = "no",
  co = 0,
  dic = "",
  mecabrc = "",
  etc = ""
)
}
\arguments{
\item{charVec}{A character vector.}

\item{pos}{Parts of speech that should be extracted.
If \code{NULL}, all terms are extracted.}

\item{minFreq}{Minimum document frequency for filtering terms.
Terms that appear less than \code{minFreq} within a document are ignored.}

\item{weight}{Method to weight term frequencies.}

\item{co}{If \code{1}, returns co-occurrence matrix.}

\item{dic}{Path to a user dictionary file such as \code{ishida.dic}.}

\item{mecabrc}{Path to a mecabrc file.}

\item{etc}{Other options for 'MeCab' tagger.}
}
\value{
An integer matrix is invisibly returned.
}
\description{
Creates a document-term matrix out of a character vector.
Each cell of the matrix shows the actual frequency of each word.
}
