% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docNgramDF.R
\name{docNgramDF}
\alias{docNgramDF}
\title{docNgramDF}
\usage{
docNgramDF(
  mojiVec = "MeCab",
  type = 0,
  pos = "Default",
  baseform = 0,
  minFreq = 1,
  N = 1,
  kigo = 0,
  weight = "no",
  co = 0,
  dic = "",
  mecabrc = "",
  etc = ""
)
}
\arguments{
\item{mojiVec}{A character vector.}

\item{type}{Kind of tokens. \code{0} for character, \code{1} for term.
Defaults to \code{0}.}

\item{pos}{Parts of speech that should be extracted.
If \code{NULL}, all terms are extracted.}

\item{baseform}{Genkei. See \code{\link[=docDF]{docDF()}}. Defaults to \code{0}.}

\item{minFreq}{Minimum document frequency for filtering terms.
Terms that appear less than \code{minFreq} within a document are ignored.}

\item{N}{Unit of tokens. If \code{2}, counts bi-grams.}

\item{kigo}{If \code{1}, \verb{[[TOTAL-TOKENS]]} includes number of symbols.
Defaults to \code{0} (does not count symbols).}

\item{weight}{Method to weight term frequencies.}

\item{co}{If \code{1}, returns co-occurrence matrix.}

\item{dic}{Path to a user dictionary file such as \code{ishida.dic}.}

\item{mecabrc}{Path to a mecab resource file.}

\item{etc}{Other options for 'MeCab' tagger.}
}
\value{
A data frame.
}
\description{
Creates a data.frame of N-grams out of a character vector.
}
