% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ngram.R
\name{Ngram}
\alias{Ngram}
\title{Ngram}
\usage{
Ngram(
  filename,
  type = 0,
  N = 2,
  pos = "Default",
  dic = "",
  mecabrc = "",
  etc = ""
)
}
\arguments{
\item{filename}{An input file.}

\item{type}{Kind of tokens. \code{0} for character, \code{1} for term.
Defaults to \code{0}.}

\item{N}{Unit of tokens. If \code{2}, counts bi-grams.}

\item{pos}{Parts of speech that should be extracted.
If \code{NULL}, all terms are extracted.}

\item{dic}{Path to a user dictionary file such as \code{ishida.dic}.}

\item{mecabrc}{Path to a mecab resource file.}

\item{etc}{Other options for 'MeCab' tagger.}
}
\value{
A data.frame.
}
\description{
Returns a data.frame of N-gram.
}
\examples{
if (anyRcfileExists()) {
  text_file <- system.file("samples/doc1.txt", package = "RMeCab")
  out <- Ngram(text_file, type = 1)
  head(out)
}
}
