% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NgramDF2.R
\name{NgramDF2}
\alias{NgramDF2}
\title{NgramDF2}
\usage{
NgramDF2(
  directory,
  type = 0,
  pos = "Default",
  minFreq = 1,
  N = 2,
  kigo = 0,
  dic = "",
  mecabrc = "",
  etc = ""
)
}
\arguments{
\item{directory}{A directory in which text files are stored
or a single file.}

\item{type}{Kind of tokens. \code{0} for character, \code{1} for term.
Defaults to \code{0}.}

\item{pos}{Parts of speech that should be extracted.
If \code{NULL}, all terms are extracted.}

\item{minFreq}{Minimum document frequency for filtering terms.
Terms that appear less than \code{minFreq} within a document are ignored.}

\item{N}{Unit of tokens. If \code{2}, counts bi-grams.}

\item{kigo}{If \code{1}, \verb{[[TOTAL-TOKENS]]} includes number of symbols.
Defaults to \code{0} (does not count symbols).}

\item{dic}{Path to a user dictionary file such as \code{ishida.dic}.}

\item{mecabrc}{Path to a mecabrc file.}

\item{etc}{Other options for 'MeCab' tagger.}
}
\value{
A data.frame is invisibly returned.
}
\description{
Creates a data.frame of N-grams out of all files in a given directory.
}
\examples{
\dontrun{
text_dir <- system.file("samples", package = "RMeCab")
out <- NgramDF2(text_dir, type = 1)
head(out)
}
}
