% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{rmecab-args-tagger}
\alias{rmecab-args-tagger}
\alias{rmecab-args-set}
\alias{rmecab-args-weight}
\alias{rmecab-args-kigo}
\alias{rmecab-args-pos}
\alias{rmecab-args-minFreq}
\alias{rmecab-args-co}
\alias{rmecab-args-mypref}
\title{Arguments for tagger}
\arguments{
\item{dic}{Path to a user dictionary file such as \code{ishida.dic}.}

\item{mecabrc}{Path to a mecab resource file.}

\item{etc}{Other options for 'MeCab' tagger.}

\item{type}{Kind of tokens. \code{0} for character, \code{1} for term.
Defaults to \code{0}.}

\item{N}{Unit of tokens. If \code{2}, counts bi-grams.}

\item{weight}{Method to weight term frequencies.}

\item{kigo}{If \code{1}, \verb{[[TOTAL-TOKENS]]} includes number of symbols.
Defaults to \code{0} (does not count symbols).}

\item{pos}{Parts of speech that should be extracted.
If \code{NULL}, all terms are extracted.}

\item{minFreq}{Minimum document frequency for filtering terms.
Terms that appear less than \code{minFreq} within a document are ignored.}

\item{co}{If \code{1}, returns co-occurrence matrix.}

\item{mypref}{If \code{1}, returns basic forms of terms.}
}
\description{
Arguments for tagger

Arguments set type and N

Argument weight

Argument kigo

Argument pos

Argument minFreq

Argument co

Argument mypref
}
\keyword{internal}
