% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{RMeCab-package}
\alias{RMeCab}
\alias{RMeCab-package}
\title{RMeCab: Interface to 'MeCab'}
\description{
Parses Japanese texts with 'MeCab'. The original 'MeCab' is licensed under the BSD 3-Clause "New" or "Revised" License. See the "LICENSE.note" file for its license notice.
}
\references{
\url{http://www.amazon.co.jp/dp/4627848420}
}
\seealso{
\url{https://rmecab.jp/wiki/index.php?RMeCab}
}
\author{
\strong{Maintainer}: Motohiro Ishida \email{ishida.motohiro@gmail.com}

Other contributors:
\itemize{
  \item Taku Kudo [copyright holder]
  \item Nippon Telegraph and Telephone Corporation [copyright holder]
}

}
\keyword{internal}
