% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMeCabC.R
\name{RMeCabC}
\alias{RMeCabC}
\title{RMeCabC}
\usage{
RMeCabC(str, mypref = 0, dic = "", mecabrc = "", etc = "")
}
\arguments{
\item{str}{A string scalar to be tokenized.}

\item{mypref}{If \code{1}, returns basic forms of terms.}

\item{dic}{Path to a user dictionary file such as \code{ishida.dic}.}

\item{mecabrc}{Path to a mecab resource file.}

\item{etc}{Other options for 'MeCab' tagger.}
}
\value{
A list.
}
\description{
Takes a string as an argument and tokenize it into
a length-1 lists of term.
}
