% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMeCabDF.R
\name{RMeCabDF}
\alias{RMeCabDF}
\title{RMeCabDF}
\usage{
RMeCabDF(dataf, coln, mypref = 0, dic = "", mecabrc = "", etc = "")
}
\arguments{
\item{dataf}{A data.frame.}

\item{coln}{Column number or name which include the text to analyze.}

\item{mypref}{If \code{1}, returns basic forms of terms.}

\item{dic}{Path to a user dictionary file such as \code{ishida.dic}.}

\item{mecabrc}{Path to a mecab resource file.}

\item{etc}{Other options for 'MeCab' tagger.}
}
\value{
A list.
}
\description{
Takes a data frame as an argument and tokenize it into
a length-1 lists of term.
}
\details{
This is a wrapper of \code{\link[=RMeCabC]{RMeCabC()}}.
Any blanks should be replaced with \code{NA} for \code{coln}.
}
