% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMeCabDoc.R
\name{RMeCabDoc}
\alias{RMeCabDoc}
\title{RMeCabDoc}
\usage{
RMeCabDoc(filename, mypref = 1, kigo = 0, dic = "", mecabrc = "", etc = "")
}
\arguments{
\item{filename}{An input file.}

\item{mypref}{If \code{1}, returns basic form of terms.}

\item{kigo}{If \code{1}, \verb{[[TOTAL-TOKENS]]} includes number of symbols.
Defaults to \code{0} (does not count symbols).}

\item{dic}{Path to a user dictionary file such as \code{ishida.dic}.}

\item{mecabrc}{Path to a mecabrc file.}

\item{etc}{Other options for 'MeCab' tagger.}
}
\value{
A list.
}
\description{
Takes a file as an argument and tokenize it into
a list of term.
}
\examples{
\dontrun{
text_file <- system.file("samples/doc1.txt", package = "RMeCab")
unlist(RMeCabDoc(text_file))
}
}
