% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMeCabFreq.R
\name{RMeCabFreq}
\alias{RMeCabFreq}
\title{RMeCabFreq}
\usage{
RMeCabFreq(filename, dic = "", mecabrc = "", etc = "")
}
\arguments{
\item{filename}{an input file.}

\item{dic}{Path to a user dictionary file such as \code{ishida.dic}.}

\item{mecabrc}{Path to a mecab resource file.}

\item{etc}{Other options for 'MeCab' tagger.}
}
\value{
A data.frame.
}
\description{
Takes text files as first argument
and returns parts of speech and frequencies as a data.frame.
}
\examples{
\dontrun{
text_file <- system.file("samples/doc1.txt", package = "RMeCab")
RMeCabFreq(text_file)
}
}
