% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMeCabText.R
\name{RMeCabText}
\alias{RMeCabText}
\title{RMeCabText}
\usage{
RMeCabText(filename, dic = "", mecabrc = "", etc = "")
}
\arguments{
\item{filename}{An input file}

\item{dic}{Path to a user dictionary file such as \code{ishida.dic}.}

\item{mecabrc}{Path to a mecabrc file.}

\item{etc}{Other options for 'MeCab' tagger.}
}
\value{
A list.
}
\description{
Takes a file as an argument and tokenize it into
a list of terms and parts of speech.
}
\examples{
\dontrun{
text_file <- system.file("samples/doc1.txt", package = "RMeCab")
RMeCabText(text_file)
}
}
