% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOptChr.R
\name{anyRcfileExists}
\alias{anyRcfileExists}
\title{Check if any mecabrc file exists}
\usage{
anyRcfileExists()
}
\value{
A logical.
}
\description{
Check if any mecabrc file exists
}
\details{
This is a helper function that checks if any mecabrc file exists
before initializing tagger.

MeCab expects a mecabrc file to be present;
if not, it will raise an error (without any message!).
}
