% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collocate.R
\name{collocate}
\alias{collocate}
\title{collocate}
\usage{
collocate(filename, node, span = 3, dic = "", mecabrc = "", etc = "")
}
\arguments{
\item{filename}{An input file.}

\item{node}{Node word.}

\item{span}{Windows span. Defaults to \code{3}.}

\item{dic}{Path to a user dictionary file such as \code{ishida.dic}.}

\item{mecabrc}{Path to a mecab resource file.}

\item{etc}{Other options for 'MeCab' tagger.}
}
\value{
A data.frame.
}
\description{
Finds collocations from the specified text file.
Takes a \code{node} word and a window \code{span} as arguments.
}
\examples{
if (anyRcfileExists()) {
  text_file <- system.file("samples/doc1.txt", package = "RMeCab")
  out <- collocate(text_file, "\u6570\u5b66")
}
}
