% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docNgram.R
\name{docNgram}
\alias{docNgram}
\title{docNgram}
\usage{
docNgram(
  mydir,
  type = 1,
  N = 2,
  pos = "Default",
  dic = "",
  mecabrc = "",
  etc = ""
)
}
\arguments{
\item{mydir}{A directory where text files are stored.}

\item{type}{Kind of tokens. \code{0} for character, \code{1} for term.
Defaults to \code{0}.}

\item{N}{Unit of tokens. If \code{2}, counts bi-grams.}

\item{pos}{Parts of speech that should be extracted.
If \code{NULL}, all terms are extracted.}

\item{dic}{Path to a user dictionary file such as \code{ishida.dic}.}

\item{mecabrc}{Path to a mecab resource file.}

\item{etc}{Other options for 'MeCab' tagger.}
}
\value{
A data.frame is invisibly returned.
}
\description{
Creates a data.frame of N-gram out of all files in a given directory.
}
\examples{
\dontrun{
text_dir <- system.file("samples", package = "RMeCab")
out <- docNgram(text_dir, type = 1)
head(out)
}
}
