% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{RMeCab-package}
\alias{RMeCab}
\alias{RMeCab-package}
\title{RMeCab: Interface to 'MeCab'}
\description{
Parses Japanese texts with 'MeCab'.
}
\references{
\url{http://www.amazon.co.jp/dp/4627848420}
}
\seealso{
\url{https://rmecab.jp/wiki/index.php?RMeCab}
}
\keyword{internal}
