\name{rmSign}
\alias{rmSign}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ rmSign }
\description{
 \code{ rmSign} signs from the result of analyze. 
% Japanese-sentence default value has been set up at 256 characters.  If
% longer sentences are included, the forth argument can be raised up to 102400.
}
\usage{
%  rmSign(str, mypref, mybuf)
  rmSign(x)
% 
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ RMeCab object}
%  \item{mybuf}{ enough buffer for japanese strings}
}
%% \details{
%%   If necessary, more details than the description above 
%% }
\value{
  returns a data.frame.
}
\references{  \url{http://www.amazon.co.jp/dp/4627848420}   }
\author{ Motohiro ISHIDA \email{ishida.motohiro@gmail.com} }%\author{ishida motohiro }
% \note{ note
%  \section{Warning }{....} 
% }
% \seealso{ objects to See Also as \code{\link{help}},  }
% \examples{
% %   \dontrun{
% %   }
% }
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ misc }
% \keyword{  }% __ONLY ONE__ keyword per line
